third_party/blink/web_tests/fast/multicol/vertical-lr/break-properties-expected.txt

PASS: 'break-before' is in column 2
PASS: 'after-break' is in column 4
PASS: 'no-break' is in column 5
